changes to Netlify-functions-continue-after-completion.md
Hello Serge,
Your blog was very helpful and well done, as with the rest of your blogs. 
I messed around with the setTimeout and modern and legacy styles extensively today, and discovered the following changes that I propose to the blog.
Indeed both the legacy and modern styles work when running the function with the netlify-cli locally, but once deployed only the modern style now works.

Cheers.